[DEBT-043-D17] Split research/thesis.py into focused modules
- Extract ThesisStatus, ThesisEvidence, Thesis to _thesis_models.py (253 lines)
- Keep ThesisTracker in thesis.py (177 lines)
- Re-exports preserve backwards-compatible public API
- Quality gates: pre-commit, mypy, pytest (1052 passed)
- Acceptance criteria: 1/1 complete
